// src/lib/api/models/bodyCreateWasteReport.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * FastAPI JWT Auth Boilerplate
 * FastAPI Boilerplate with JWT Authentication
 * OpenAPI spec version: 0.1.0
 */
import type { BodyCreateWasteReportLatitude } from './bodyCreateWasteReportLatitude';
import type { BodyCreateWasteReportLongitude } from './bodyCreateWasteReportLongitude';
import type { WasteType } from './wasteType';
import type { UrgencyLevel } from './urgencyLevel';
import type { BodyCreateWasteReportDescription } from './bodyCreateWasteReportDescription';
import type { BodyCreateWasteReportReporterPhone } from './bodyCreateWasteReportReporterPhone';

export interface BodyCreateWasteReport {
  /** @maxLength 255 */
  street_address: string;
  /** @maxLength 100 */
  city: string;
  /** @maxLength 20 */
  postal_code: string;
  latitude?: BodyCreateWasteReportLatitude;
  longitude?: BodyCreateWasteReportLongitude;
  waste_type: WasteType;
  quantity_estimate: string;
  urgency_level: UrgencyLevel;
  description?: BodyCreateWasteReportDescription;
  /** @maxLength 100 */
  reporter_name: string;
  reporter_phone?: BodyCreateWasteReportReporterPhone;
  photos?: Blob[];
}