// src/lib/api/models/wasteType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * FastAPI JWT Auth Boilerplate
 * FastAPI Boilerplate with JWT Authentication
 * OpenAPI spec version: 0.1.0
 */

export type WasteType = typeof WasteType[keyof typeof WasteType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WasteType = {
  household: 'household',
  recyclable: 'recyclable',
  electronic: 'electronic',
  construction: 'construction',
  hazardous: 'hazardous',
  organic: 'organic',
  illegal_dumping: 'illegal_dumping',
  other: 'other',
} as const;